refactor: extract runtime summarization policy
